src/services/textsense-ocr/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
Pillow==10.2.0
requests==2.31.0
python-multipart==0.0.6
numpy==1.26.4
paddlepaddle==2.6.1
paddleocr==2.7.3